% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lyric_analysis.R
\name{lyric_analysis}
\alias{lyric_analysis}
\title{lyric_analysis}
\usage{
lyric_analysis(
  artist,
  dictionary = "GI",
  duplicates = FALSE,
  start_year = 0,
  end_year = as.numeric(format(Sys.Date(), "\%Y")),
  exact = FALSE
)
}
\arguments{
\item{artist}{The musical artist of interest (case insensitive)}

\item{dictionary}{The type of dictionary to use in the sentiment analysis (one of "GI" or "QDAP")}

\item{duplicates}{True/False. Whether to keep duplicate entries of songs. Default is false. (specifically for songs that were hits for more than one year)}

\item{start_year}{The beginning of the date range to analyze, including this value (default is 0, but data starts at 1960)}

\item{end_year}{The end of the date range to analyze, including this value (default is the current year, but data goes up to 2016 currently)}

\item{exact}{True/False. Whether to look for an exact artist match or to use the artist as a prefix. (default is false. if exact, may not return songs that feature other artists)}
}
\value{
A data frame with names and sentiment values of any top 100 songs from your specified artist
}
\description{
This function takes in an artist name and returns a dataframe of the song names and lyric sentiment values for that artist's Billboard Top 100 hits.
}
\examples{
lyric_analysis(artist = "Britney Spears", dictionary = "QDAP", duplicates = "FALSE", start_year = 1990, end_year = 2010, exact = FALSE) # with maximum specifications
lyric_analysis(artist = "Madonna") # with minimum specifications
}
\author{
Lauren Harris & Sarah Lam
}
