% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_artists.R
\docType{data}
\name{all_artists}
\alias{all_artists}
\title{All Artists}
\format{
A vector of strings.
\itemize{
\item Artist name
}
}
\source{
\url{https://cran.r-project.org/package=billboard}
}
\usage{
all_artists
}
\description{
Contains all unique artists in 'billboard' data for a user to search artists function will run on. Includes features as unique artists.
}
\keyword{datasets}
